Modify improper augeas set command
This is a syntax fix that properly sets the global_filter as the
resulting quoting was off. (I had thought that running lvmconfig --validate
would properly catch such an error, but it must not inspect the type.) In any
case, I've run lvmdiskscan to confirm that normal LVM operations are working
properly with the new configuration.

Change-Id: Icf3b95c70abe2d5f1f9c69fa868cc7a9f35f5379