Verify that the Deployment Plan creation was successful
The creation of the default deployment plan was recently broken but the error
went unnoticed as CI never failed. This then led to another obscure bug as
something worked due only to the default plan not existing.

This change waits for the plan creation to finish and verifies that it
completed without an error. A timeout is added, should, for some reason the
creation ever stall.

NOTE: This will not raise an error if plan creation fails, since it is
      currently broken. This will be fixed in this following patch:
      https://review.openstack.org/373446

Partial-Bug: #1622683
Change-Id: Id8001fee677fe321b42892264e3c55afb6790e1c