Ensure that the default plan was created successfully
In change Id8001fee677fe321b42892264e3c55afb6790e1c we wait for
the workflow to finish, but don't verify that it didn't error as
there is an issue with the default plan creation. There is a fix
for this which this patch depends on.

This change also expands on the error messages so users can more
easily debug the problem.

Closes-Bug: #1622683
Depends-On: Ic494ed0874dd79e148ef9c7093bdcf32bed4d6ae
Change-Id: Iaaf528372b57c19480c47fe22213b27b8ca871a7