Use same logging format for file and stream
The file logging handler was using a timestamped output format,
while the stream handler was just dumping the log message without
any extra context.  This creates some weird issues if you want to
be explicit that a message is an error, because prepending ERROR:
to the logging string results in a double message in the log file.

Since the timestamped logging format is probably useful anyway,
let's just use that everywhere and not have to worry about
redundantly specifying the severity of a log message.  This also
allows us to remove the awk pipe for the puppet apply since we'll
already be timestamping the output.

Change-Id: I2bafc5f167eea3dbac77d2720c1df0672d1276e3