Switch ironic-inspector to mysql
Using sqlite is legacy and is no longer recommended. It won't work with the planned
HA architecture as well. This change switched to a new MySQL database.

Note that this has a big upgrade impact: introspection information will no longer
be available. I'm not sure how to deal with it, but it won't break too much,
assuming that users don't run introspection right during the upgrade.

Change-Id: I5f029139867ff3e7eb3d140f72e452dc7412ecdc
Depends-On: Ib4d3548e76a9314592d8b66707df1987d7cfc5b9