Use kwargs when creating mistralclient
Recently in If6ff35e91c3d35c2741332c7e739bb92b1234c54 a refactor of the
authentication in Mistral client stopped it accepting *args and now it
only accepts **kwargs.

kwargs were always expected, but Python will allow you to implicitly
pass them based on the order, but as this was removed and now **kwargs
is in the method signature it wont work.

This is a regression in mistralclient.

Closes-Bug: #1628853
Change-Id: I8b8963c97b9e8c2a31b5f554f737d95685bc330a