Switch to use Puppet Openstack spec helper
Rather than continuing to manage the Gemfile, we can leverage the Puppet
Openstack spec helper which will keep our gem dependencies in line with
the other puppet modules.  By leveraging the Puppet Openstack spec
helper we can leverage the dependency fixes that are usually found in
the other modules first.

Change-Id: I6fee3c45ea135d53341df6145b33f62205f82161