Make the snmpd_readonly_user_password available to Mistral
The Mistral deployment workflow needs to access this password so it can
be passed to Heat as the SnmpdReadonlyUserPassword parameter. Otherwise
the workflow relies on being passed the password, for tripleoclient this
isn't an issue but TripleO-ui doesn't have access to this.

The password is stored in a Mistral environment, which can then be accessed
by Mistral actions.

Partial-Bug: #1631279
Change-Id: I71edae4a4dee2204edf32e3b2800f075d221b856
(cherry picked from commit 0c5b729609c687f0fcadaafb69b56d9363b8d9b6)