Tune number of workers for undercloud
By default many services use the number of cpus as the default number of
workers.  On baremetal systems this can lead to excessive RAM
comsumption as this can be 16, 32, 48 or higher for each of the
services. In the upstream puppet openstack, a fact was created as part
of the Newton cycle that is the larger value between (number of cpus/4)
or 2 but is capped at 8. These defaults have not been updated yet in the
upstream modules and it would be better to specifically call out the
worker configurations in our puppet-stack-config.yaml template.

This change updates the worker settings for the classes that are pulled
in for the undercloud to leverage this tuned fact to lower the overall
number of workers for all the services.  This should improve memory
consumption and reduce the number of processes running on baremetal
systems. The os_workers fact will be 2 for unless the cpu count is
greater than 8 with an incremental increase of 1 worker for every 4
processors until 32 processors.  Anything above 32 processors will still
result in a worker value of 8.

This change updates 17 services that currently may use the number of
processors for the default number of workers.

Change-Id: I9ed855648e23b0a7e452e6a840a92779fa3f6d48
Closes-Bug: #1633544