Make module name and file name equivalent
Newer checkmodule commands have a requirement that the selinux module
name and file name are equivalent. This patch updates the ipxe element
to match that requirement.

This patch is applied directly to mitaka since the ipxe element was
removed in newton. Upon merging the plan is to apply to liberty as well.

Change-Id: Id01e8084a238d939bbef2c82e1baa82812dc64b3
Closes-Bug: #1635030