Install, enable docker service
This change installs and enables the docker service when
enable_docker_registry=true. This will allow the undercloud docker
registry to be populated with images that are built on the undercloud
or pulled from other registries.

The docker service runs on a local socket so there are no networking
concerns with enabling the service.

A dormant docker service takes about 500MB of memory, so enabling it
should have minimal impact on the undercloud memory use.

Change-Id: I22854dfa735027d7916bc0090432b66250f7780f