Add code to support novajoin in the undercloud
Novajoin is a new nova metadata micro-service that registers
instances created by nova with a FreeIPA server, so that we can
take advantage of server/service identities, kerberos credentials,
access and sudo controls etc.

The included modules will cause the undercloud to be registered with
a FreeIPA server, for novajoin to be installed, and for nova on the
undercloud to be configured to use novajoin.

Implements: blueprint novajoin
Depends-On: I5ffa45bdc400e123079c79e15776ebacdcb24de9
Change-Id: Ia2fdb544460cd39fbfda5970c9bfba22feec6403