Increase the default number of workers for heat engine
We switched to a saner default for service workers on the undercloud
with this patch[1]. However, reducing the number of workers for heat
engine is not so sane. The massive nested stack we are deploying grinds
to a halt with only two engine workers, which is now our saner default
for any system with 8 or fewer CPU cores.

This patch changes to using the heat default for the number of engine
workers which is max(#CPUs,4).

[1] https://git.openstack.org/cgit/openstack/instack-undercloud/commit/?id=ee5f25a726910961caf72de2c6c55de06c922b74

Change-Id: I95df0f39f37316cc56eafb351c823197f252d7b7
Related-Bug: 1638908