Validate vips when generating certificate too
When generate_service_certificate is True,
undercloud_service_certificate will not necessarily be set when it
is passed to validation.  We need to check if either value is set
when deciding whether to validate vips.

Unit tests for this behavior were missing as well, so those have
been added.

Another consideration for this change is that we have started
passing non-IP values to these vip parameters when configuring
ssl.  This is counterintuitive, but apparently works as intended
so let's just rename the parameters and handle both IPs and DNS
names for those values.

Change-Id: I53151d4f555d5d161a3e53ce5f022e3bf3b2ffbd
Closes-Bug: 1643655