Add option to set undercloud dns nameserver
This adds an undercloud_nameserver option which will be taken into
account in os-net-config to set up the undercloud's nameserver. If left
unset, it will give an empty list to the dns_servers parameter in
os-net-config, which will actually do nothing. If set, it will use that
nameserver as the sole member of a list that is passed to the
aforementioned parameter.

Co-Authored-By: Juan Antonio Osorio Robles <jaosorior@redhat.com>
Change-Id: I2b8a3f64dc1958f00b0471b68ba8283da76a4e7b