Optional Cinder support for undercloud
This is required to support overcloud nodes boot from volume. This support is
newly introduced in ironic in the Ocata, see the following specs for details:
http://specs.openstack.org/openstack/ironic-specs/specs/approved/boot-from-volume-reference-drivers.html
http://specs.openstack.org/openstack/ironic-specs/specs/approved/volume-connection-information.html

The implementation is mostly copied from overcloud cinder services, except
that only iscsi backend is enabled, as ironic only supports it for now.

The backup service is not enabled as part of this patch for simplicity.

Partially implements: blueprint undercloud-boot-from-volume

Change-Id: I7068c85be421d7642514c30bc47bc05589d43551