Improve upgrade process to include upgrade flag
Previously, the upgrade process for the undercloud has been to just
rerun the install process. The problem is that this does not include
anwyay to indicate that it's already been run if we need to trigger some
different actions on subsequent runs.  As part of the Newton to Ocata
upgrade, we need to be able to run the cell v2 setup in a different
order than on a traditional install. We need a way for the puppet
scripts that ultimately get run to know if it's an upgrade or install
action. This change adds the creation of an undercloud_upgrade fact that
will be used when the upgrade process is run. Additionally a new script
called instack-upgrade-undercloud has been created to be used when
upgrade. The use of instack-install-undercloud will cause the install
flow to be execuated, while the instack-upgrade-undercloud will set the
fact to true to allow for the puppet scripts to use the upgrade flow.

Change-Id: Ie3cb21e30334fe8ffc0a9d6e707b42269b64c9ec
Related-Bug: #1649341
Related-Blueprint: undercloud-upgrade