Add additional proxy and config endpoints for UI
Include additional hiera variables in the format of
<scheme>://<host>:<port> suitable for use in both Apache mod_proxy rules
and UI configuration file generation.  Two sets of hiera variables are
created, since they vary slightly depending on the context in which
they're referenced.   This work also includes fixup code for the proper
format of a Websocket URL appropriate for tunnelling wss through https.

Tests were also modified to reflect the current number of endpoints.

This change is required by other work being done to the rest of the
associated blueprint, against the tripleo-puppet repo.

Change-Id: Ib20f4b0891563ae90ec80675635a64c39bd2fdb7
Implements: blueprint proxy-undercloud-api-services