Remove setup-neutron
We don't really need to use setup-neutron from os-cloud-config to create
the ctlplane network and subnet. As simple as they are, it's more
straightforward and clear to just use neutron commands directly.

This removes all depedencies on os-cloud-config from instack-undercloud
so we can remove that requirement from requirements.txt (and moves us
closer to deprecating that entire repo).

Change-Id: I88a4359884429f85f1ef356d50abb0bf664a5540