Fix initialization of novaclient
projectid argument of novaclient's(< 7.0) entry-point had several meaning
in case of different cases. It is not a user-friendly behaviour, so it was
fixed in 7.0 . Now projectid means project/tenant id in terms of keystone,
like it should be from the beginning.

tenant/project name should be transmitted viaa project_name or tenant_name
keyword argument.

Closes-Bug: #1654183
Change-Id: I106ee603e0853bbc2da4b99724e83587de3cb4ba