Stop deploying Nova API in WSGI with Apache
It was suggested by Nova team to not deploying Nova API in WSGI with
Apache in production.
It's causing some issues that we didn't catch until now (see in the bug
report). Until we figure out what was wrong, let's disable it so we can
move forward in the upgrade process.

Note: once it's supported by Nova, we'll revert this patch.
Change-Id: I2712ca0b9626771cec1f3d98b04cc8c18eb1cb15
Related-Bug: 1661360