Clear any previous environment in stackrc
As was done in I6badc5864426a371323a59100155e1c284af84b8 for the
overcloud.  This is particularly a problem because right now we
only generate a v2 stackrc file, and if you switch from a v3
overcloudrc to the v2 stackrc you'll have v3 env vars left over in
the environment, which will likely break client usage.

Given that it's late in the cycle, I think this is the safest
change to make.  Adding v3 settings to stackrc can be a followup
in Pike.

Change-Id: I2ba3160656372fd73d51cfbe9de8f49dc983009d
Closes-Bug: 1661689
(cherry picked from commit 03d5cd9a17525ea44dd2c69d2d95b75c3b7f4c0d)