Increase sync timeout for nova db syncs
We have seen on lower quality hardware that the nova db syncs can take
an excessive amount of time. In order to still support deploying on this
hardware, let's increase the timeout from the default 300 seconds to 900
seconds to allow for this less performant gear. It is not recommended to
increase this past 900 as if we start hitting this then we need to be
understanding what is occuring in these db syncs. 300 seconds should be
enough time to setup a database especially on install. But there are
cases for upgrades or slower disks where it can exceed 300 seconds.

Change-Id: I77507c638237072e38d9888aff3da884aeff0b59
Related-Bug: #1660722
Related-Bug: #1661396