Change _member_role_exists to work with current upgrade flow
Because we now shut down all OpenStack services before upgrading
the undercloud, we can't run API queries against them prior to
running puppet.  This means we now need to handle the _member_ role
in a post config step.

Note that this was only not failing before because the check in
_stackrc_exists was always returning false, so the old code was
never actually run.  That also means it was not doing what it was
supposed to, of course.

Change-Id: Iaf81adc48321dbc985c50aeb5bfc0cc94810cb1d
Closes-Bug: 1663719