Include swap in memory check
It's not uncommon for developers to use SSD-backed swap to allow
for more instances on a limited memory baremetal system.  Previously
this was not allowed by the memory check because it only looked at
physical memory.  This change includes swap in the memory check so
that use case is supported out of the box.

Change-Id: If7fdf189fadd39cd6380c15a35f4b7d49cc42a96