Respect OS_LOG_CAPTURE env var
OS_LOG_CAPTURE is used by the test logging fixture to determine
whether to swallow log output so it isn't printed to the console
during unit tests.  This cleans up the test output considerably,
but because we set up a log handler ourselves in this project we
are side-stepping that functionality and printing log output anyway.

This change looks for the OS_LOG_CAPTURE env var and skips the
StreamHandler setup that is causing the undesired output if it is
found.

Change-Id: Ie2276313064cb805ebf218ff164f5fe83e330223