Use purged for firewalld
If the undercloud node was installed via anaconda gui, the anaconda-core
and related packages may be installed. These packages have a dependency
on firewalld which can result in an install failure. To ensure that we
don't fail, we need to use purge which will also cause the dependencies
to get removed as well.

Change-Id: Ib46e1545d3843ff0306e6711939840fe2fb407dc
Closes-Bug: #1669915