Disable docker-registry by default
The docker-registry package has been replaced by docker-distribution in
centos and is failing the puppet code that expect a docker-registry
service. A proper patch I89e14cc2a27299ce4c191d2a823deb0424693831 was
submitted to stable/newton and stable/ocata but was a bit too invasive
for mitaka.

Simply changing the default not to install docker-registry should make
the gate happy again.

Change-Id: I9f5599889ac3e56ba201b01340f21fc7e4051df5