Add auth/authtoken configuration for novajoin
novajoin has usually just used the nova user for running and even file
permissions; however, as nova now supports passing the keystone token
for the vendordata plugin, and as novajoin should support being run in a
node (or container) where nova is not available, it makes sense to start
having an own user for this vendordata plugin service.

Thus, this commit adds that.

Change-Id: I3e0da54c4191745f4acb880c608e5b20ac06b914
Depends-On: I190a84a5aaf1fcc301f0605931b24d5de6999a8b