Configurable [clients]/endpoint_type for heat.conf
Makes the [clients]/endpoint_type in heat.conf configurable instead of
hardcoded to internal. The hieradata_override configuration could be
used to set the hiera value.

This ought to be configurable in the case where using deployed servers
and the ctlplane is not routable from the deployed servers network(s).
In that situation, you would have to use the undercloud's public
endpoints (with an ssl hostname set via undercloud_public_host) for Heat
polling and signaling.

Depends-On: If707c5623c0e34e9219eeafdafaf0ac42daf5c8d
Closes-Bug: #1674822
Co-Authorized-By: James Slagle <jslagle@redhat.com>
Co-Authorized-By: Emilien Macchi <emilien@redhat.com>

Change-Id: I8102392ddceda0a55be6767cbbda572ac2fda0b9