Create Heat API CFN endpoint
Create the Heat API CFN endpoint by including the needed puppet class.
Previously, we were not creating the endpoint.

When using deployed servers where the ctlplane is not routable, you may
want to configure Heat polling and signaling over the public endpoints
instead of the internal endpoints. In order to do so, endpoints for the
CFN service need to be created.

Change-Id: Iefa9d5bd3b03eb4812f24be2ca1ad0fa3a157402
Closes-Bug: #1674827