Set default domain for all keystone users
The resource Keystone::Resource::Service_identity gets a parameter
called default_domain. This parameter comes into effect if no user or
project domain have been specified. So it's a good way to get all the
kesytone services to use this domain. The aforementioned resource is in
charge of creating the relevant keystone users, roles, services and
endpoints for OpenStack services.

Change-Id: Ic2e6fe9503a0af5fd652a5cbf25262bd705e3951