Add authentication parameters for novajoin vendordata plugin
With nova's vendordata plugin supporting authentication, we need to
specify the authentication parameters on the metadata side. This will be
used to get a keystone token, and will subsequently be validated by
novajoin's middleware, where keystone authtoken middleware is. This way
we can asure that log that requests actually come from an authenticated
source (nova).

Change-Id: I78bb217c485bad342e7ce8d60786fe1e5acf55a4
Depends-On: I870e244aff97439143b58e9b2284830b4388cc5f