Fix broken command prompt
Interactive editing of the command line was broken due to the usage of
$PROMPT_COMMAND. The cursor would often be placed at the wrong location
when editing history due to prompt length confusion. We already had the
OS_CLOUDNAME variable for this purpose, and I took the prompt setting
logic from what we previously used in tripleo-incubator:

http://git.openstack.org/cgit/openstack/tripleo-incubator/tree/cloudprompt

A previous attempt to fix this bug is in
I55b8a9222fc53f33371bae0baa3c0e93340672b9
but it seems there is some disagreement on the approach.

Change-Id: Iae1d0e942475ebace38c2afd24b08cece14c931d
Partial-Bug: #1673459