Plan format migration: keep the Mistral env around
Until the tripleo-common change at [1] is merged, operations related to
Plans still expect a Mistral environment to exist. Deleting it now is
causing updates and re-deployments to fail after upgrading the
undercloud (which also migrates the Plan to use the new format).

Let's keep the Mistral environment around until the tripleo-common
changes are in and fix the TODO then.

[1] https://review.openstack.org/#/c/452291/6

Related blueprint: stop-using-mistral-env
Closes-Bug: #1685171

Change-Id: Icac6efe17fffc78bf74a34206cced82e2b968fcf