undercloud/upgrade: change ownership for ironic-dbsync.log
A change was done in puppet-ironic:
https://review.openstack.org/#/c/457478/

It broke TripleO upgrades because ironic-dbsync.log used to be owned by
root:root and now it's managed by ironic user in puppet-ironic, which is
good.

Because we want to support existing deployments, we need a way to handle
the new owership change, so let's do it in a upgrade task in
undercloud.py. Puppet will be able to run ironic-dbsync and we'll remove
this workaround after Pike release.

Change-Id: I325c30f49b765ef6413743491902e97735877321