Allow setting the DNS domain name in undercloud.conf
Without being able to do this, users who want to set CloudDomain
during their deployment will have to manually override neutron's
dns_domain config, which either requires custom hieradata or the
reapplication of the customization after each undercloud update.
This is not a good user experience, so we should allow them to set
it once in the config file and never have to touch it again.

It turns out that the Nova dhcp_domain option must also be set,
even though it is a deprecated option.  A bug with nova has been
opened to address this in the future.

Change-Id: I61e5fc005aa27ba76f91a2b84fec751af93061b5
Related-Bug: 1698010