Move UI package install to base manifest
We moved the package definition in
tripleo::profile::base::ui
to tripleo::ui.

So forth, this path needs to be updated.

This is used in I5cce963a6225552f4b7253cb34edc72b392b9dda
as we removed tripleo::profile::base::ui
and the package install is now in tripleo::ui

Change-Id: Ie20ecabea91ca4c2040c5ef3bf6c71b2b53d26ef
Closes-bug: 1698105
Depends-on: I5cce963a6225552f4b7253cb34edc72b392b9dda