Revert "Default neutron's dns_domain to undercloud's domain"
Per my comment on the original change, I have concerns about
this change to the default behavior of the undercloud.  I
haven't gotten a response to my comment there (probably because
the patch is already merged), so I'm proposing that we revert
it until there is a clear path forward for making the defaults
in the overcloud templates work sanely with this.

This reverts commit 437b116e461396da1a7e94f1c6afde36cb380dae.

Change-Id: Id8e8cfd3d82fad6e31a4224a7897e932f0690ddb