Fail-fast if the undercloud-passwords.conf file is missing
On undercloud updates/upgrades, if the undercloud-passwords.conf
file was deleted or moved it will break some services because there
are passwords that cannot be changed.  This validates that the file
exists and stops immediately with a message explaining the problem
if the file is missing.

Change-Id: Ic83d6afc8dc2128874b97cc414ea28272e7d6639
Closes-Bug: 1702709
(cherry picked from commit 74123de347437e6953429f27ab1a89adcf8ac085)