Remove isolated-build element
The element is no longer used in the undercloud install as it's not
contained in any of the json files under json-files or included by any
element dependencies.

This is a fix for the security issue where the /tmp path is used.

Change-Id: Ib8013fa33cd14d7e4a66e07bd6f3a280c41a7f15
Closes-Bug: #1712380
(cherry picked from commit 722d9e4292f2013c4c13ce391778b902c88d83f1)