Validate local_interface exists
In order to improve the user experiance, we should validate that the
provided local_interface actually exists prior to trying to configure
the undercloud with it. If net_config_override is not used and the
local_interface does not exist, an error will be returned to the
user early on in the installation process. Additionally if the user
does not have an undercloud.conf, a warning is printed to show that
the defaults will be used.

Change-Id: Ia2a7e7fc916f6d2b62e212cb42641494b88d6d17
Closes-Bug: #1709177