Disable compute auto disabling
As part of Pike, nova introduced a change to have the nova-compute
process automatically disable the nova-compute instance in the case of
consecutive build failures. This can lead to odd errors when deploying
the ironic nodes on the undercloud as you end up with a ComputeFilter
error. This change disables this functionality for the undercloud since
we do not want the nova-compute instance running on the undercloud for
Ironic to be disabled in the case of multiple deployment failures.

Change-Id: Ia8a4cfcd6b31b496161cba14ee597bc61af2cab4
Depends-On: If46602fdcac38745e6b6b17d560844bb4f42ba3c
Closes-Bug: #1712632