Enable convergence engine on the undercloud
We want to turn this on for Queens, and we might as well do it early
in the cycle so we have a chance to address any issues that may come
up.

We also want to migrate any existing stacks to use the convergence
engine to reduce the support matrix.  This patch includes a post-config
step do so.

Change-Id: I96952f78cb95252d2bc6c0df64561d07df51dc5a