Only delete workflows with the tripleo-common-managed tag
The new workflow tag allows us to detect tripleo-common workflows. It is
safe to delete and reload these. Workflows that come from other places
shouldn't be removed.

There is some upgrade logic here that handles the O -> P upgrade, as on
the first run the workflows will not have the tripleo-common-managed
tag. If the tag isn't found, we revert to the current behaviour and then
on future undercloud installations this wont be a problem.

The new test is a duplication of a previous test but with multiple
workflows that have tripleo in the name. Only the one with the tag is
deleted.

Depends-On: Ieeb51f8e705b8988fb4d0f5d17d7fce6b5b04f36
Closes-Bug: #1715389
Change-Id: I84d441daedcafe686b9576d47b037977b493dbc7
(cherry picked from commit d6c0636421d2186690a4f0d46533593f4ad8ad19)