Generate a standalone stackrc
Previously the stackrc file made a hiera call to get the admin
password.  The problem with this is that it makes the stackrc file
non-portable to any system except the undercloud.  Writing the
actual password into stackrc will allow it to work on any system
that has network connectivity to the undercloud.

Note that this does not have any negative impact on the security
of the stackrc file since previously the user sourcing it would have
had to have sudo access to read the password from hiera anyway.  If
anything this is more secure because it does not require allowing
root access to any user of stackrc.

Depends-On: I6ccb832b52a476b5fa659a9a8f09e38ae41d324b
Change-Id: I7a6b665cf9d40889e8e72475887c7ce166e77a40