Set notification_bind on zaqar websocket
Don't let zaqar websocket listen to notifications on all interfaces:
otherwise, it uses the hostname as a subscription target, and it doesn't
work when the host is set as an IPv6 host. Ultimately this should be
fixed in Zaqar, but this will work around it for now.

Change-Id: I78e0671024b93ec64740cd7a8cda227a5145b789
Related-Bug: #1727406
(cherry picked from commit 5fdb4d7913bfc0096c725f6f99fd98e949c75d58)