Load undercloud DB password to a mistral environment
This patch adds a new variable to the mistral
environment 'tripleo.undercloud-config' allowing
to retrieve the root password for the MySQL
databases. Now this variable can be accessed from any
Mistral workflow.
Also, this patch adds the ability of checking if
the variables are not loaded properly due to
an upgrade, in that case we update the Mistral
environment to have all the variables in place.

Change-Id: I822c3d96b5805909656d8e92e963b851f3bd77de
(cherry picked from commit be3bd48214e6719db9bd364f020ea5b652b9a536)
Related-Bug: 1742191