Increase the timeout for plan creation from 6 mins to 9
Uploading the Heat templates to swift can take over 4 minutes. This only
leaves 2 for all the other stages of a plan creation. We should make
this faster but until that happens we should increase the timeout to
reduce CI failures.

See the related bug for a larger description of the problem.

Related-Bug: #1738493
Change-Id: I67411cc5bbec1ad0b6a6efa11a2c5a811c135e70