Set password for mysql root user on undercloud
The port is firewalled off by default so mysql is not reachable except
to localhost users. In any case it is good to set a password for the
mysql root user.

We are 'reusing' the user_db_password parameter, because it is actually
unused, instead of adding yet another parameter.

Tested this change against an existing undercloud and observed that
the mysql root user got a correct password and that the /root/.my.cnf
file got generated.

Closes-Bug: #1742191

Change-Id: I408ce3a0fe2ab8e86bcc280256cdb51688efde75
(cherry picked from commit 955c339af3ddf040fab4a153f1e6c249aefe0989)