Mariadb online upgrade after yum update
During undercloud upgrade, yum can update the mariadb-server package.
If that is the case and the updated package is a major mariadb update
(e.g. 5.5 -> 10.1, or 10.0 -> 10.1), some database system tables might
need to be updated.

Ensure that mysql_upgrade will be run after a major mariadb update.
Also, force-stop the mariadb server before the upgrade to workaround
a rpm scriptlet limitation when going 5.5 -> 10.1.

Change-Id: Iebc3f39cb2dae355124ba39f7653ef8d20baacc6