Increase size of Heat max_template_size
Most of the HTTP servers default max body size to 1MB, Apache, Nginx.
That default value is ok as long as users mostly get data from server
but when something needs to be uploaded, 1MB limit is often not enough.
This is what we see with Heat and some of the large templates.
It's not uncommon that WSGI apps have 10MB+ body size limit,
so with the current Heat default value we are still in low range sizes.

Original Heat 'max_template_size' default value was 1M.
In instack-undercloud we bumped it to 2M, bug #1667697.
This change increases it to 4MB.

Change-Id: Ic913ecb44dac800131292d1f3661be5b71f35a78
Closes-Bug: 1752467