Make user-provided cert take priority over autogenerated one
It used to be the case that if you give both a user-provided
certificate, and set the flag for instack to autogenerate the
certificate, the autogenerated one took precedence. This is not ideal
since it might not be what the user expects (especially if we switch to
using TLS by default).

Closes-Bug: #1755497
Change-Id: Ia8aa52d80999ad278501ca8ecf0638ef8de6ca19