Fix next_hop for metadata service host route on local_subnet
The host route to the metadata service should use the
undercloud as next_hop when configuring the subnet that
is on the same L2 segment as the undercloud itself. The
code used to manage this was not correctly implemented.

Change-Id: Icf40af8ac424d14386aacda38cf674cc25273fc8
Closes-Bug: #1757418
(cherry picked from commit 73d8effd15d45c28ecababbbca2787c6d152bff8)