Always include certmonger_user
This makes sure we always include the certmonger_user profile, which
installs the local CA (if that's the one we're using). This is necessary
for when we deploy TLS by default in the overcloud.

It also makes the setting of the certificate specification to be
optional and to depend on the generate_service_certificate flag.

Change-Id: I8b46ce3f9cd6e36d0b8f604b49e4113301461a4c