Fix net_config_override validations
There is an issue with the _validate_no_ip_change function
where it doesn't work correctly with a custom net_config_override
file.

This patch resolves validation issue so that custom os-net-config
templates are taken into account. Additionally it makes it so
that the net_config_override can be empty, thus providing the
user a means of disabling os-net-config if they choose to do so.

Change-Id: Iada099a6bddc7927d26b19776e705b1ac231b2e5
Depends-On: I69eb812af2d653cb3eb29694ae29b57ebb2de53d
Closes-bug: #1764507