Make sure we start nova-compute after ironic-conductor.
We need to ensure that ironic-conductor starts before nova-compute.
This is to workaround an issue where nova-compute tries and fails to
call plug_vifs, this in turn report a vm_state error which, in this
case is a false positive.  See lp#1777608 for more.

We ensure ordering by forcing puppet to restart nova-compute after
ironic-conductor in the case of undercloud upgrade/update.

Change-Id: Ifbada53f088258a397777a6fa18dd7c1b37c09d3
Closes-Bug: #1777475