Drop example in config file comments
There have been some reports of users misinterpreting
where to place configuration for the provisining subnets
in undercloud.conf. The example text for the subnets leads
to users adding the subnets section directly after the
subnets option. I.e in the middle of the DEFAULT section
of the ini file.

This change drops the example. The ctlplane-subnet section
is there as an example already.

Related-Bug: rhbz1585497
Related-Bug: #1778779
Change-Id: I417623db410330058d30fe7435112f01a0f72c83