Update stackrc to v3 auth values when --use-heat is false.
During pike we have updated the deployment to use v3 authentication in
stackrc.  But we didn't upgrade the existing stackrc.

We didn't use it in the code so it was not a problem.

But in queens, when we use '--use-heat' is false, we check the status
of the overcloud before starting and we load auth url from stackrc and
force domain in the api request.  So we have a v2 url and a v3
request. That fails.

Now, we make sure that the stackrc is in v3 version before starting
the undercloud upgrade in queens when we use --use-heat set to false.

Change-Id: I3f008b9466408fe797ca50ebf96651a72fc9a544
Closes-Bug: #1798553