[queens-only] include local_ip & undercloud_admin_host in docker_insecure_registries
This patch includes by default the local_ip and undercloud_admin_host in
docker_insecure_registries parameter, that can be missed but actually
needed in some cases.

This code was done in tripleoclient for the containerized undercloud in
Rocky, so we just need to port it here for stable versions, to have the
same behavior.

Change-Id: I03be8381ec776fd769139c79b85470acedc64d2b
Closes-Bug: #1799784