Get the stack name if possible before running the validations group
If we run the group validations in the Undercloud and we
do not specify the stack name, this will fail if the
Overcloud stack is not called 'overcloud'.

In this case we try to fetch the stack name and we compare it
with the swift containers to match the deployment plan name
if there is a match, then we know the stack name.

Change-Id: Ib2e27b1b43f573385c089e57bd2290daac49d814
Closes-Bug: 1806684
Resolves: rhbz#1573120