Enable override of overcloud status check for undercloud update.
We need to migrate heat to a new engine during pike to queen upgrade
and that requires an overcloud in a sane state.

The problem is that during undercloud update (from queen to latest
queen) we may need to update the undercloud even if the overcloud is
in a failed state.  For instance if we need a fix in a mistral
workbook to make an overcloud action successful.

This patch offers the possibility to by-pass the overcloud check in
such case.

A simple environment variable setting is used to communicate with the
python script triggered by instack-pre-upgrade-undercloud et
instack-upgrade-undercloud. It looks cleaner that mangling with the
python -c line in the scripts.

So if the environment variable TRIPLEO_FORCED_UPDATE is set to
non-empty string or if TRIPLEO_FORCED_UPDATE is passed as an argument
of the two scripts mentioned above then we trigger the no-check heat
stack status mode.

Change-Id: I43a1b4892f25bb649ee34d39b5fa462a302834aa
Partial-Bug: #1806683